requirements.txt
pytest
xmltodict
selenium
webdriver_manager
chromedriver
requests
allure-pytest
pytest-rerunfailures
pytest-xdist